Move LambdaToMethod at the end of the compiler pipeline
