+++ src/jdk.compiler/share/classes/com/sun/tools/javac/tree/JCTree.java
@@ -2013,6 +2013,7 @@ public abstract class JCTree implements Tree, Cloneable, DiagnosticPosition {
         public JCTree body;
         public boolean canCompleteNormally = true;
         public ParameterKind paramKind;
+        public JCExpression methodReceiverExpression;
 
         public JCLambda(List<JCVariableDecl> params,
                         JCTree body) {
